chore: Move trade API to coordinator
