+++ coordinator/src/lib.rs
@@ -1,13 +1,14 @@
-pub mod cli;
-pub mod logger;
-pub mod orderbook;
-pub mod routes;
-pub mod schema;
-
 use diesel::PgConnection;
 use diesel_migrations::embed_migrations;
 use diesel_migrations::EmbeddedMigrations;
 use diesel_migrations::MigrationHarness;
+
+pub mod cli;
+pub mod logger;
+pub mod node;
+pub mod orderbook;
+pub mod routes;
+pub mod schema;
 
 pub const MIGRATIONS: EmbeddedMigrations = embed_migrations!();
